Revert "Move v1 to /v1/ route, add redirect"
