+++ package.json
@@ -27,8 +27,7 @@
     "gh-pages": "^6.1.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-redux": "^9.1.0",
-    "react-router-dom": "^6.26.1"
+    "react-redux": "^9.1.0"
   },
   "devDependencies": {
     "@testing-library/dom": "^9.3.4",
