+++ src/App.tsx
@@ -37,8 +37,7 @@ const App = () => {
         </Grid>
         <footer>
           <Typography variant="caption">
-            Found a bug? Report it on{" "}
-            <a href="https://github.com/Nubuscu/panic-pwa/issues">Github</a>
+            Found a bug? Report it on <a href="https://github.com/Nubuscu/panic-pwa/issues">Github</a>
           </Typography>
         </footer>
       </Container>
