+++ src/app/components/CurrentHero.tsx
