+++ src/app/components/character/Archetype.tsx
