+++ src/app/components/character/Style.tsx
