+++ src/app/components/config/BossConfig.tsx
