+++ src/app/components/config/StoogeConfig.tsx
