+++ src/app/components/config/WarriorConfig.tsx
