+++ src/app/components/config/hero/HeroConfig.tsx
