+++ src/app/components/config/hero/focusedStances.tsx
