+++ src/app/components/config/hero/franticStances.tsx
