+++ src/app/components/config/hero/fusedStances.tsx
