+++ src/app/components/config/hero/stances.css
