+++ src/app/textContent.ts
